Restructure the components directory
