+++ src/components/hud/TimeIndicator.tsx
